+++ CHANGES.md
@@ -1,7 +1,7 @@
 Change Log
 ----------------
 - Version 0.9.3 (Pending)
-    
+    - Select the number of badges to generate when printing test badges
 
 - Version 0.9.2 (7/21/2016)
     - Added back/forward button support to Check In by Badge report